pyproject.toml
[project]
name = "opnsense-dyndns-hetzner"
version = "0.2.0"
description = "Dynamic DNS updater for Hetzner Cloud DNS using OPNsense WAN interface IPs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hcloud>=2.13",
    "dnspython>=2.6",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.22",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
odh = "opnsense_dyndns_hetzner.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
